backend/requirements.txt

fastapi
uvicorn
azure-identity
azure-ai-generative
python-multipart
